Created a goal reached function and added verbose
Added a function to check if we've reached our goal. Also added a
verbose option in order to allow for printing out data while the GA is
running. Moved our problem specific function to the new file
defaultGAProblems, not everything has been moved here that should be,
but the most obvious have been.