Renamed Chromosome to Organism
Also renamed chr.genes in organism to chromosome, this fits better with
biological model. We still have Organism$values as organism is an
environment which holds a list of organisms. It might make more sense if
population held a list rather than an env, but that's how the full
system is setup.