Added spatial reproduction for coevolution
We now reproduce with spatial reproduction for coevolution. We deal with
the fitness issue by basing fitness in both populations by the previous
generation's population. We have also dealt with some elitism for
spatial GA's by doing a hard tournament between that node and the elite
that was there in the previous generation.

We also have added single child crossover where only the parent node is
replaced. Will need to add more customization options for this.

Currently a bug with how we display top fitness, we only check a single
population's fitnesses. Should probably display fitnesses for each
population.

With our spatial population we should also change how 'rest' is
selected, just select the rest of the population.