Changed crossover levels and fitness reporting
Updated the crossover level to ignore the number of elites. This clearly
doesn't make much sense in a local elitism perspective, and it also
doesn't effect the full level GA so much as we generally run a low level
of elites with respect to the population size.

Also updating the fitness tracking to only record the relevant results.